pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Devin AI <158243242+devin-ai-integration[bot]@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.121.0"}
psycopg = {extras = ["binary"], version = "^3.2.12"}
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.20"
pydantic-settings = "^2.11.0"
httpx = "^0.28.1"
redis = "^7.0.1"
scikit-learn = "^1.7.2"
xgboost = "^3.1.1"
prophet = "^1.2.1"
pandas = "^2.3.3"
numpy = "^2.3.4"
stripe = "^13.2.0"
python-dotenv = "^1.2.1"
bcrypt = "^5.0.0"
sqlalchemy = "^2.0.44"
psycopg2-binary = "^2.9.11"
alembic = "^1.17.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
